New semantics for functions
When a function is instantiated (becoming a "closure"), it is stored
in the memory and replaced by its address in the program. This
makes values in the language more regular; all values are "simple
stuff": numbers, addresses, nil, or boxed versions of those. In turn,
this greatly simplifies using any value as a table index.

This commit also eliminates 'BaseType', which was isomorphic to
'IRType'.